Finishing Home/About page
Revamp overall look over previous iteration. Focus on simplicity and to the point presentation of Home page.
Home page is mostly done (around 80%).
Added shortcut on bottom left to GroundLeaf wiki. Which is reused into "return to Home" when applied to "Practice" section. This might change in the future, seems cluttery for what its worth.